Fourth Iteration
The favourites function now works, although limited. It seems to work best for Chrome browsers.

If you press the star button on either Brazil or Greece, it will be added to favourites.html. 'Delete all' will clear LocalStorage. Due to time constraints, we haven't been able to get all buttons work in a similar way, but that should be relatively easy to do.